Classpath changes and Linux testing
Introduced environment variables
* SCALA_HOME (must contain scala-library.jar version 2.9.1)
* WEKA_HOME  (must contain weka.jar version 3.7.5)
* JODA_HOME  (must contain joda-time-2.0.jar)
* KNOWING_HOME (must contain
de.lmu.ifi.dbs.knowing.core_0.1.0.201111241009.jar)
* VLC_HOME (must contain vljc-1.2.2.jar and jna.jar)

Upgraded to Java Version 7

### Problems under Ubuntu
Found some older versions (1.1.2) of the VLC libraries  (libvlc.so ,
libvlccore)
which makes it possible to start the programm. However with some
Exceptions
(old versions).

With the correct version 1.2 I'm not able to start the application.
This is because libm.so.6 couldn't be loaded. However it exists. In
_EvalStart.java_ I try to load these libraries before the application
starts to get more verbose exceptions.